Change request retry method.
If a PipeModule resolves an issue of a request and, at the same time,
it is related to authorization, it is necessary to recreate
a HTTP provider to apply new authorization rules. This can apply
not only to the authorization procedure but to everything that modifies
a request this way or another.